refactor: cleaned up file structure and imports
