+++ frontend/components/Inventory/CreatePartForm.js
@@ -4,9 +4,9 @@ import { Button } from "@mui/material";
 import TextField from "@mui/material/TextField";
 import Autocomplete from "@mui/material/Autocomplete";
 import SendIcon from "@mui/icons-material/Send";
-import Box from "@mui/material";
 
-export default function UpdateSupplierForm() {
+
+export default function CreatePartForm() {
   const [productValue, setProductValue] = React.useState("");
   const [productId, setProductId] = React.useState("");
