+++ frontend/components/Inventory/CreateProductForm.js
@@ -2,11 +2,7 @@ import * as React from "react";
 import { useEffect, useState } from "react";
 import { Button } from "@mui/material";
 import TextField from "@mui/material/TextField";
-import Autocomplete from "@mui/material/Autocomplete";
 import SendIcon from "@mui/icons-material/Send";
-import Box from "@mui/material";
-
-//GET LINK: http://localhost:8788/productInventory/parts
 
 export default function CreateProductForm() {
   const [inputValue, setInputValue] = React.useState("");
