+++ frontend/components/Procurement/Procurements.js
@@ -9,39 +9,25 @@ import List from "@mui/material/List";
 import Typography from "@mui/material/Typography";
 import Divider from "@mui/material/Divider";
 import IconButton from "@mui/material/IconButton";
-import Badge from "@mui/material/Badge";
 import Container from "@mui/material/Container";
 import Grid from "@mui/material/Grid";
 import Paper from "@mui/material/Paper";
 import Link from "@mui/material/Link";
 import MenuIcon from "@mui/icons-material/Menu";
 import ChevronLeftIcon from "@mui/icons-material/ChevronLeft";
-import NotificationsIcon from "@mui/icons-material/Notifications";
-import PedalBikeIcon from "@mui/icons-material/PedalBike";
-import { mainListItems, secondaryListItems } from "./listItems";
-import Chart from "./Chart";
-import Deposits from "./Deposits";
+import { mainListItems, secondaryListItems } from "../listItems";
 import Copyright from "../Copyright";
 import SettingsIcon from "@mui/icons-material/Settings";
 import Menu from "@mui/material/Menu";
 import MenuItem from "@mui/material/MenuItem";
-import Title from "./Title";
-import BasicForm from "../Dashboard/BasicForm";
-import ContactsForm from "./ContactsForm";
+import Title from "../Title";
 import UpdateSupplierForm from "./UpdateSupplierForm"
 import UpdateContactForm from "./UpdateContactForm"
 import CreateSupplierForm from "./CreateSupplierForm"
 import CreateContact from "./CreateContact"
 import RemoveSupplierForm from "./RemoveSupplierForm"
-
-import TextField from "@mui/material/TextField";
-
-import SendIcon from '@mui/icons-material/Send';
-
 import SuppliersTable from "./SuppliersTable";
-import { Button } from "@mui/material";
-
-import ContactsTable from "../Dashboard/ContactsTable";
+import ContactsTable from "./ContactsTable";
 
 const drawerWidth = 240;
