+++ frontend/components/Sale/CreateStoreForm.js
@@ -2,9 +2,8 @@ import * as React from "react";
 import { useEffect, useState } from "react";
 import { Button } from "@mui/material";
 import TextField from "@mui/material/TextField";
-import Autocomplete from "@mui/material/Autocomplete";
 import SendIcon from "@mui/icons-material/Send";
-import Box from "@mui/material";
+
 
 export default function CreateStoreForm() {
   const [inputValue, setInputValue] = React.useState("");
