+++ frontend/components/Sale/CreateStoreSale.js
@@ -4,7 +4,6 @@ import { Button } from "@mui/material";
 import TextField from "@mui/material/TextField";
 import Autocomplete from "@mui/material/Autocomplete";
 import SendIcon from "@mui/icons-material/Send";
-import Box from "@mui/material";
 
 export default function CreateStoreSale() {
   const [productValue, setProductValue] = React.useState("");
