+++ frontend/components/Sale/OnlineSalesTable.js
@@ -1,9 +1,8 @@
 import * as React from "react";
 import { useEffect, useState } from "react";
 import { DataGrid } from "@mui/x-data-grid";
-import Title from "./Title";
 
-export default function StoresTable() {
+export default function OnlineSalesTable() {
   const [keyword, setKeyword] = useState("sales");
   const [data, setData] = useState([]);
   const fetchData = () => {
@@ -12,7 +11,6 @@ export default function StoresTable() {
       .then((data) => setData(data._embedded.onlineSaleList))
       .catch((err) => console.error(err));
   };
-  
 
   useEffect(() => {
     fetchData();
@@ -22,11 +20,9 @@ export default function StoresTable() {
     { field: "id", headerName: "ID", width: 100, minWidth: 75, maxWidth: 200 },
     { field: "customerName", headerName: "Customer Name", width: 125, minWidth: 150, maxWidth: 200 },
     { field: "customerAddress", headerName: "Customer Address", width: 250, minWidth: 200, maxWidth: 600 },
-
     { field: "productName", headerName: "Product Name", width: 125, minWidth: 150, maxWidth: 200 },
     { field: "quantity", headerName: "Quantity", width: 100, minWidth: 50, maxWidth: 200},
     { field: "orderStatus", headerName: "Order Status", width: 200, minWidth: 150, maxWidth: 200 }
-    
   ];
 
   const rows = [];
@@ -42,13 +38,11 @@ export default function StoresTable() {
 
   return (
     <div style={{ height: 400, width: "100%" }}>
-      {/* <Title>Suppliers</Title> */}
       <DataGrid
         rows={rows}
         columns={columns}
         pageSize={5}
         rowsPerPageOptions={[5]}
-        // checkboxSelection
       />
     </div>
   );
