+++ frontend/components/Sale/Sale.js
@@ -16,22 +16,13 @@ import Paper from "@mui/material/Paper";
 import Link from "@mui/material/Link";
 import MenuIcon from "@mui/icons-material/Menu";
 import ChevronLeftIcon from "@mui/icons-material/ChevronLeft";
-import NotificationsIcon from "@mui/icons-material/Notifications";
-import PedalBikeIcon from "@mui/icons-material/PedalBike";
 import { mainListItems, secondaryListItems } from "./listItems";
-import Chart from "./Chart";
-import Deposits from "./Deposits";
 import Copyright from "../Copyright";
 import SettingsIcon from "@mui/icons-material/Settings";
 import Menu from "@mui/material/Menu";
 import MenuItem from "@mui/material/MenuItem";
-import Title from "./Title";
-import UpdatePartForm from "./UpdatePartForm";
-import UpdateProductForm from "./UpdateProductForm";
-import CreatePartForm from "./CreatePartForm";
-import ProductsTable from "./ProductsTable";
+import Title from "../Title";
 import StoresTable from "./StoresTable";
-
 import CreateStoreForm from "./CreateStoreForm";
 import CreateStoreSale from "./CreateStoreSale";
 import CreateOnlineSale from "./CreateOnlineSale";
