+++ frontend/pages/inventory/index.js
@@ -1,4 +1,3 @@
-import BasicForm from "../../components/Dashboard/BasicForm";
 import Inventory from "../../components/Inventory/Inventory";
 
 export default function inventory() {
