+++ frontend/pages/procurement/index.js
@@ -1,4 +1,3 @@
-import BasicForm from "../../components/Dashboard/BasicForm";
 import Procure from "../../components/Procurement/Procurements";
 
 export default function procurement() {
