+++ frontend/pages/sale/index.js
@@ -1,4 +1,3 @@
-import BasicForm from "../../components/Dashboard/BasicForm";
 import Sale from "../../components/Sale/Sale";
 
 export default function sale() {
